requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
frozenlist==1.4.0
idna==3.4
joblib==1.3.1
magic-filter==1.0.9
multidict==6.0.4
nltk==3.8.1
numpy==1.25.1
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
telebot==0.0.5
threadpoolctl==3.2.0
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
